requirements.txt
streamlit==1.32.0
pandas==2.1.4
plotly==5.18.0
python-binance==1.0.19
numpy==1.26.3
python-dotenv==1.0.1
